+++ MbDCode/DiagonalMatrix.h
@@ -9,24 +9,24 @@ namespace MbD {
 	{
 		//
 	public:
-		DiagonalMatrix(size_t count) : Array<T>(count) {}
+		DiagonalMatrix(int count) : Array<T>(count) {}
 		DiagonalMatrix(std::initializer_list<T> list) : Array<T>{ list } {}
-		void atiputDiagonalMatrix(size_t i, std::shared_ptr < DiagonalMatrix<T>> diagMat);
+		void atiputDiagonalMatrix(int i, std::shared_ptr < DiagonalMatrix<T>> diagMat);
 		std::shared_ptr<FullColumn<T>> timesFullColumn(std::shared_ptr<FullColumn<T>> fullCol);
-		size_t nrow() {
-			return this->size();
+		int nrow() {
+			return (int) this->size();
 		}
-		size_t ncol() {
-			return this->size();
+		int ncol() {
+			return (int) this->size();
 		}
 		double sumOfSquares() override;
-		size_t numberOfElements() override;
+		int numberOfElements() override;
 		void zeroSelf() override;
 	};
 	template<typename T>
-	inline void DiagonalMatrix<T>::atiputDiagonalMatrix(size_t i, std::shared_ptr<DiagonalMatrix<T>> diagMat)
+	inline void DiagonalMatrix<T>::atiputDiagonalMatrix(int i, std::shared_ptr<DiagonalMatrix<T>> diagMat)
 	{
-		for (size_t ii = 0; ii < diagMat->size(); ii++)
+		for (int ii = 0; ii < diagMat->size(); ii++)
 		{
 			this->at(i + ii) = diagMat->at(ii);
 		}
@@ -36,9 +36,9 @@ namespace MbD {
 	{
 		//"a*b = a(i,j)b(j) sum j."
 
-		auto nrow = this->size();
+		auto nrow = (int) this->size();
 		auto answer = std::make_shared<FullColumn<T>>(nrow);
-		for (size_t i = 0; i < nrow; i++)
+		for (int i = 0; i < nrow; i++)
 		{
 			answer->at(i) = this->at(i) * fullCol->at(i);
 		}
@@ -48,7 +48,7 @@ namespace MbD {
 	inline double DiagonalMatrix<double>::sumOfSquares()
 	{
 		double sum = 0.0;
-		for (size_t i = 0; i < this->size(); i++)
+		for (int i = 0; i < this->size(); i++)
 		{
 			double element = this->at(i);
 			sum += element * element;
@@ -56,15 +56,15 @@ namespace MbD {
 		return sum;
 	}
 	template<typename T>
-	inline size_t DiagonalMatrix<T>::numberOfElements()
+	inline int DiagonalMatrix<T>::numberOfElements()
 	{
-		auto n = this->size();
+		auto n = (int) this->size();
 		return n * n;
 	}
 	template<>
 	inline void DiagonalMatrix<double>::zeroSelf()
 	{
-		for (size_t i = 0; i < this->size(); i++) {
+		for (int i = 0; i < this->size(); i++) {
 			this->at(i) = 0.0;;
 		}
 	}
