+++ MbDCode/DirectionCosineIeqctJeqc.h
@@ -8,7 +8,7 @@ namespace MbD {
         //pAijIeJept ppAijIeJepEIpt ppAijIeJepEJpt ppAijIeJeptpt 
     public:
         DirectionCosineIeqctJeqc();
-        DirectionCosineIeqctJeqc(EndFrmcptr frmi, EndFrmcptr frmj, size_t axisi, size_t axisj);
+        DirectionCosineIeqctJeqc(EndFrmcptr frmi, EndFrmcptr frmj, int axisi, int axisj);
         void initialize();
         void initializeGlobally();
         void calcPostDynCorrectorIteration() override;
