+++ MbDCode/DispCompIeqcJecO.cpp
@@ -7,7 +7,7 @@ MbD::DispCompIeqcJecO::DispCompIeqcJecO()
 {
 }
 
-MbD::DispCompIeqcJecO::DispCompIeqcJecO(EndFrmcptr frmi, EndFrmcptr frmj, size_t axis) : DispCompIecJecO(frmi, frmj, axis)
+MbD::DispCompIeqcJecO::DispCompIeqcJecO(EndFrmcptr frmi, EndFrmcptr frmj, int axis) : DispCompIecJecO(frmi, frmj, axis)
 {
 }
