+++ MbDCode/DispCompIeqcJeqcKeqct.h
@@ -8,7 +8,7 @@ namespace MbD {
         //priIeJeKept ppriIeJeKepXIpt ppriIeJeKepEIpt ppriIeJeKepXJpt ppriIeJeKepEJpt ppriIeJeKepEKpt ppriIeJeKeptpt 
     public:
         DispCompIeqcJeqcKeqct();
-        DispCompIeqcJeqcKeqct(EndFrmcptr frmi, EndFrmcptr frmj, EndFrmcptr frmk, size_t axisk);
+        DispCompIeqcJeqcKeqct(EndFrmcptr frmi, EndFrmcptr frmj, EndFrmcptr frmk, int axisk);
         void initialize() override;
         void calcPostDynCorrectorIteration() override;
