+++ MbDCode/DispCompIeqctJeqcO.cpp
@@ -7,7 +7,7 @@ MbD::DispCompIeqctJeqcO::DispCompIeqctJeqcO()
 {
 }
 
-MbD::DispCompIeqctJeqcO::DispCompIeqctJeqcO(EndFrmcptr frmi, EndFrmcptr frmj, size_t axis) : DispCompIeqcJeqcO(frmi, frmj, axis)
+MbD::DispCompIeqctJeqcO::DispCompIeqctJeqcO(EndFrmcptr frmi, EndFrmcptr frmj, int axis) : DispCompIeqcJeqcO(frmi, frmj, axis)
 {
 }
