+++ MbDCode/GEFullMat.h
@@ -8,7 +8,7 @@ namespace MbD {
         //
     public:
         std::shared_ptr<FullMatrix<double>> matrixA;
-        void forwardEliminateWithPivot(size_t p) override;
+        void forwardEliminateWithPivot(int p) override;
         void backSubstituteIntoDU() override;
         void postSolve() override;
         void preSolvewithsaveOriginal(SpMatDsptr spMat, FColDsptr fullCol, bool saveOriginal) override;
