+++ MbDCode/GESpMatFullPv.h
@@ -1,20 +1,20 @@
 #pragma once
 
-#include "GESpMat3.h"
+#include "GESpMat.h"
 
 namespace MbD {
     class GESpMatFullPv : public GESpMat
     {
         //positionsOfOriginalCols rowPositionsOfNonZerosInColumns 
     public:
-        void doPivoting(size_t p) override;
-        void forwardEliminateWithPivot(size_t p) override;
+        void doPivoting(int p) override;
+        void forwardEliminateWithPivot(int p) override;
         void backSubstituteIntoDU() override;
         void postSolve() override;
         void preSolvewithsaveOriginal(SpMatDsptr spMat, FColDsptr fullCol, bool saveOriginal) override;
 
-        std::shared_ptr<std::vector<size_t>> positionsOfOriginalCols;
-        std::shared_ptr<std::vector<std::shared_ptr<std::vector<size_t>>>> rowPositionsOfNonZerosInColumns;
+        std::shared_ptr<std::vector<int>> positionsOfOriginalCols;
+        std::shared_ptr<std::vector<std::shared_ptr<std::vector<int>>>> rowPositionsOfNonZerosInColumns;
     };
 }
