+++ MbDCode/GESpMatFullPvPosIC.h
@@ -10,11 +10,11 @@ namespace MbD {
         //system pivotRowLimits pivotRowLimit 
     public:
         void preSolvewithsaveOriginal(SpMatDsptr spMat, FColDsptr fullCol, bool saveOriginal) override;
-        void doPivoting(size_t p) override;
+        void doPivoting(int p) override;
 
         PosICNewtonRaphson* system; //Use raw pointer when pointing backwards.
-        std::shared_ptr<std::vector<size_t>> pivotRowLimits;
-        size_t pivotRowLimit;
+        std::shared_ptr<std::vector<int>> pivotRowLimits;
+        int pivotRowLimit;
     };
 }
