+++ MbDCode/MatrixGaussElimination.cpp
@@ -5,7 +5,7 @@ using namespace MbD;
 FColDsptr MbD::MatrixGaussElimination::basicSolvewithsaveOriginal(SpMatDsptr spMat, FColDsptr fullCol, bool saveOriginal)
 {
 	this->preSolvewithsaveOriginal(spMat, fullCol, saveOriginal);
-	for (size_t p = 0; p < m; p++)
+	for (int p = 0; p < m; p++)
 	{
 		this->doPivoting(p);
 		this->forwardEliminateWithPivot(p);
