+++ MbDCode/MatrixGaussElimination.h
@@ -9,7 +9,7 @@ namespace MbD {
         //
     public:
         FColDsptr basicSolvewithsaveOriginal(SpMatDsptr spMat, FColDsptr fullCol, bool saveOriginal) override;
-        virtual void forwardEliminateWithPivot(size_t p) = 0;
+        virtual void forwardEliminateWithPivot(int p) = 0;
         virtual void backSubstituteIntoDU() = 0;
     };
 }
