+++ MbDCode/Product.cpp
@@ -19,7 +19,7 @@ Symsptr MbD::Product::differentiateWRT(Symsptr sptr, Symsptr var)
 		[var](Symsptr term) { return term->differentiateWRT(term, var); }
 	);
 	auto derivativeTerms = std::make_shared<std::vector<Symsptr>>();
-	for (size_t i = 0; i < terms->size(); i++)
+	for (int i = 0; i < terms->size(); i++)
 	{
 		auto& derivative = derivatives->at(i);
 		auto newTermFunctions = std::make_shared<std::vector<Symsptr>>(*terms);
@@ -140,7 +140,7 @@ std::ostream& MbD::Product::printOn(std::ostream& s) const
 {
 	s << "(";
 	s << *(this->terms->at(0));
-	for (size_t i = 1; i < this->terms->size(); i++)
+	for (int i = 1; i < this->terms->size(); i++)
 	{
 		s << "*" << *(this->terms->at(i));
 	}
@@ -156,6 +156,6 @@ bool MbD::Product::isProduct()
 double Product::getValue()
 {
 	double answer = 1.0;
-	for (size_t i = 0; i < terms->size(); i++) answer *= terms->at(i)->getValue();
+	for (int i = 0; i < terms->size(); i++) answer *= terms->at(i)->getValue();
 	return answer;
 }
