+++ MbDCode/SparseColumn.h
@@ -7,7 +7,7 @@ namespace MbD {
 	{
 	public:
 		SparseColumn() {}
-		SparseColumn(std::initializer_list<std::pair<const size_t, T>> list) : SparseVector<T>{ list } {}
+		SparseColumn(std::initializer_list<std::pair<const int, T>> list) : SparseVector<T>{ list } {}
 		SparseColumn(std::initializer_list<std::initializer_list<T>> list) : SparseVector<T>{ list } {}
 	};
 }
